Info/license button + copyright info at bottom
I did this early on, because it's a little bit of interactivity, but it's also quite simple - though less simple than it could have been because I wanted it to feel sleek.

I feel I've set a really good precedent for this project. It feels sleek and designed, yet retro and "developer-y".

The focus outlines, and applying focus to the back button, is needed for keyboard users. Making it lime green dashed, really helps the developer feel. Having the back button exist, and be where it is, was needed to keep the sleek design - I initially tried using a toggle button, but keeping the info/license button on screen didn't  look good in any of a few forms that came to mind.