Reverted a bit of that last commit
If DB name was long, then there wasn't always space to show it all twice.  Also if it had certain characters that we swap for html entities for security, then they see the entities in the first showing but not in the code bit, which would cause confusion and be the opposite of helpful. Showing it in that way is mostly redundant anyway, so it's better to strip it back out and leave that idea behind as impractical, rather than spend ages fixing those issues. Our users probably don't need that level of hand holding. They possibly don't need any verification at all at that stop, but showing a little bit at that point, sets a reassuring precedent.